third_party/WebKit/LayoutTests/css-parser/content-parsing-invalid.html

<!doctype html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<style>
#target1::before {
  content: normal normal;
}
#target1::after {
  content: normal normal;
}
 
#target2::before {
  content: none open-quote;
}
#target2::after {
  content: url("a") none;
}
</style>

<div id="target1"></div>
<div id="target2"></div>

<script>
test(function(){
  assert_equals(document.styleSheets[0].cssRules[0].style.content, '');
  assert_equals(document.styleSheets[0].cssRules[1].style.content, '');
  assert_equals(document.styleSheets[0].cssRules[2].style.content, '');
  assert_equals(document.styleSheets[0].cssRules[3].style.content, '');
}, "Test to ensure that invalid values of content on pseudo elements are dropped by the parser");
</script>
